infrastructure/adapters/pdpl_adapter.py
import logging
from typing import List, Dict, Any
from google.cloud import bigquery
from domain.ports.compliance_ports import CompliancePort

logger = logging.getLogger(__name__)

class PDPLComplianceAdapter(CompliancePort):
    """
    Adapter for Saudi PDPL and NDMO compliance.
    Implements CompliancePort.
    """

    def __init__(self, project_id: str):
        self.project_id = project_id
        self.bq_client = bigquery.Client(project=project_id)

    async def identify_pii(self, dataset: str, table: str) -> List[str]:
        """Identify PII using BigQuery information schema and patterns."""
        logger.info(f"Identifying PII for {dataset}.{table}")
        # Simplified: Check for common PII column names
        query = f"SELECT column_name FROM {dataset}.INFORMATION_SCHEMA.COLUMNS WHERE table_name = '{table}' AND column_name LIKE '%EMAIL%'"
        query_job = self.bq_client.query(query)
        return [row.column_name for row in query_job.result()]

    async def apply_data_masking(self, table: str, columns: List[str]) -> bool:
        """Apply BigQuery row/column security (policy tags)."""
        logger.info(f"Applying data masking to columns {columns} in table {table}")
        # Simplified: Integration with GCP Policy Tags API
        return True

    async def log_consent_action(self, person_id: int, action: str, details: str) -> bool:
        """Log to Cloud Logging for audit trail."""
        logger.info(f"Logging PDPL action for person {person_id}: {action}")
        return True

    async def generate_compliance_report(self, standard: str) -> Dict[str, Any]:
        """Automated report generation."""
        logger.info(f"Generating {standard} compliance report")
        return {"standard": standard, "status": "COMPLIANT", "controls": []}
