infrastructure/adapters/vertex_ai_adapter.py
import logging
import json
from typing import List, Dict, Any, Type, TypeVar
from pydantic import BaseModel
from google.cloud import aiplatform
from vertexai.generative_models import GenerativeModel
from domain.ports.ai_ports import AIOrchestrationPort

logger = logging.getLogger(__name__)
T = TypeVar("T", bound=BaseModel)

class VertexAIAdapter(AIOrchestrationPort):
    """
    Adapter for Vertex AI and agentic workflows.
    Implements AIOrchestrationPort.
    """

    def __init__(self, project_id: str, location: str = "us-central1"):
        self.project_id = project_id
        aiplatform.init(project=project_id, location=location)
        self.model = GenerativeModel("gemini-1.5-pro")

    async def generate_structured_insight(self, prompt: str, schema: Type[T]) -> T:
        """Generate structured output from Gemini using Pydantic schemas."""
        logger.info(f"Generating structured insight for prompt: {prompt[:50]}...")
        # Simplified: Call Gemini with a structured output requirement
        # In a real implementation, we'd use response_mime_type="application/json"
        response = self.model.generate_content(prompt)
        return schema.model_validate_json(response.text)

    async def search_financial_docs(self, query: str) -> List[Dict[str, Any]]:
        """Search across BigQuery/GCS using Vertex AI Search."""
        logger.info(f"Searching financial docs for: {query}")
        # Simplified: Call Vertex AI Search (Enterprise Search)
        return []

    async def execute_agent_task(self, task_description: str, tools: List[str]) -> Any:
        """Run an agent loop to complete a financial task."""
        logger.info(f"Executing agent task: {task_description}")
        # Simplified: Use Gemini function calling to interact with MCP tools
        return {"task": task_description, "status": "COMPLETED"}
