pyproject.toml
[project]
name = "oracleforge"
version = "0.1.0"
description = "Oracle E-Business Suite to GCP Modernization Accelerator"
authors = [
    { name = "Smeyatsky Labs", email = "info@smeyatsky.com" }
]
dependencies = [
    "python-docx>=1.2.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "dependency-injector>=4.41.0",
    "mcp>=0.1.0",
    "cx_Oracle>=8.3.0",
    "google-cloud-bigquery>=3.0.0",
    "google-cloud-storage>=2.0.0",
    "google-cloud-dataplex>=1.0.0",
    "google-cloud-datastream>=1.0.0",
    "google-cloud-aiplatform>=1.30.0",
    "asyncio>=3.4.3"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
